% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{color_hex_palettes}
\alias{color_hex_palettes}
\title{Palettes from https://www.color-hex.com}
\format{
A named list of color palettes.
}
\usage{
color_hex_palettes
}
\description{
Palettes from https://www.color-hex.com
}
\keyword{datasets}
