% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_palettes.R
\name{nn_palettes}
\alias{nn_palettes}
\title{Launch interactive Nearest neighbor finder}
\usage{
nn_palettes(palettes)
}
\arguments{
\item{palettes}{named list of palettes}
}
\value{
selected palettes
}
\description{
Launch interactive Nearest neighbor finder
}
