% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette2vec.R
\name{palette2vec}
\alias{palette2vec}
\title{Title}
\usage{
palette2vec(pals, hue_contains_n = 10)
}
\arguments{
\item{pals}{palettes}

\item{hue_contains_n}{number of hue colors used in contains columns}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Title
}
\examples{
palette2vec(list(terrain = terrain.colors(10),
                 heat = heat.colors(16),
                 topo = topo.colors(8)))
}
