% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paletteer_palettes.R
\name{paletteer_palettes}
\alias{paletteer_palettes}
\title{All discrete paletteer palettes}
\usage{
paletteer_palettes()
}
\value{
named list of color palettes
}
\description{
All discrete paletteer palettes
}
