% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt-helpers.R
\name{pals_to_gt}
\alias{pals_to_gt}
\title{Turns named list of color palettes into {gt} chart}
\usage{
pals_to_gt(pals)
}
\arguments{
\item{pals}{named list of palettes}
}
\value{
\link[gt:gt]{gt::gt} chart
}
\description{
Turns named list of color palettes into {gt} chart
}
\examples{
pals_to_gt(color_hex_palettes[1:5])
}
