% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_embedding.R
\name{umap_embedding}
\alias{umap_embedding}
\title{Launch interactive umap embedding}
\usage{
umap_embedding(palettes)
}
\arguments{
\item{palettes}{named list of palettes}
}
\value{
selected palettes
}
\description{
Launch interactive umap embedding
}
